requirements.txt
Flask
transformers
accelerate
PyPDF2
python-dotenv
